portratit only

+++ AmenLib/src/main/java/com.jaeckel/amdroid/api/model/Amen.java
@@ -45,13 +45,13 @@ public class Amen {
         this.setCreatedAt(date);
       }
 
-      this.setId(object.getInt("id"));
+      this.setId(object.getLong("id"));
 
       if (object.has("kind_id")) {
         this.setKindId(object.getInt("kind_id"));
       }
 
-      this.setUserId(object.getInt("user_id"));
+      this.setUserId(object.getLong("user_id"));
 
       User user = new User(object.getJSONObject("user"));
       this.setUser(user);
@@ -70,19 +70,19 @@ public class Amen {
     }
   }
 
-  public long getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-  public long getUserId() {
+  public Long getUserId() {
     return userId;
   }
 
-  public void setUserId(long userId) {
+  public void setUserId(Long userId) {
     this.userId = userId;
   }
